requirements.txt
discord.py[voice]
python-dotenv
ffmpeg-python